Finished creating libpychess.
The library implements the Position class to be used by the python
implementation. Currently, unfortunately, its much slower to use the C++
since converting all the types is massively expensive, and there is some
overhead imposed by the pybind11 library.

There is still scope for optimising this, however it would involve
reimplimenting Piece into C++ as well - which is not impossible and I
may come back to this at some point soon.

For now the development on this branch will be halted, since more work
on the neural network is required before publishing results. The current
state of progress is that the C++ implementation is ~1.5-2.5x slower.
Position is conditionally implemented in C++ (controlled by the USE_CPP
constant). C++ implementation of ResultKeys is a class with static int
for each of the keys, and PsuedolegalMoves(const bool colour) creates a
dict(cpp_piece, dict(int, [cpp_move])). The costly part of the
conversion comes from having to iterate through each key and perform the
actions:

1. Convert the piece from C++ piece to Py piece. It requires a
   dictionary lookup for every piece, currently done inside a dict
   comprehension. It isnt possible to just interchange the keys. Would
   be a lot more simple if libpychess::piece had the same interface and
   was used by python (in the same way as libpychess::Position).
2. Convert the std::map into a Result object. This step is actually much
   faster, since its just a dictionary with ints as the keys. This has
   required the concession that ResultKeys in python is implemented the
   same as ResultKeys in C++ - a class with static integer attributes 1
   - 5. However, it is implemented correctly and it does work well, as
     long as changes are reflected in both of the modules.

Finally, the notes of PyChessBinds will contain build instructions. Just
note that `libpychess.so` must be in the path
`vendor/pychessbinds/build/Debug|Release`.